refactor: select variable item
