+++ tests/VariablesPage.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { NextIntlClientProvider } from 'next-intl';
 import messages from '../messages/en.json';
-import VariablesPage from '@/components/pages/VariablesPage';
+import VariablesPage from '@/app/[locale]/variables/VariablesPage';
 
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn(),
